requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
flask_restful
google-cloud-logging
google-cloud-firestore